requirements.txt
Flask==2.1.0
gunicorn==20.1.0
paddlepaddle-gpu==2.1.2
insightface==0.2.1
onnxruntime